fix(ic): invalidate exact chains before reuse

+++ runtime/src/object.h
@@ -89,8 +89,11 @@ static inline void mal_object_mark_as_prototype(MalObject *object) {
     }
 }
 
-/** Current process-wide prototype-chain validity epoch (zero is never used). */
-extern u32 mal_prototype_chain_epoch;
+/**
+ * Current process-wide prototype-chain validity epoch. Zero permanently disables
+ * exact-chain caching after the theoretical u64 version space is exhausted.
+ */
+extern u64 mal_prototype_chain_epoch;
 
 /** Cold half of prototype-chain invalidation; callers use the inline flag guard. */
 void mal_object_bump_prototype_chain_epoch(void);
